Rename the file from setupStorage2.ts to setupStorage.ts
