+++ main-process/setupStorage.ts
